rename post to article
