+++ src/app/pages/home/home.component.scss
@@ -2,7 +2,7 @@
   min-height: 94%;
   /* Fallback for browsers do NOT support vh unit */
   min-height: 94vh;
-  /* 100 -6vh because the ToolBar has 6vh       */
+  /* 100 -6vh because the ToolBar has 6vh units */
 
   display: flex;
   align-items: center;
